refactor
- move all source-files to src-package (models, utilities, views and main.py)
- in app.yaml adjust path to app
