+++ list-demo-api/app.yaml
@@ -4,4 +4,4 @@ threadsafe: true
 
 handlers:
 - url: /.*
-  script: main.app
+  script: src.main.app
