+++ list-demo-api/src/models/__init__.py
@@ -1,0 +1,2 @@
+from mylist import MyList
+from myuser import MyUser
