+++ list-demo-api/src/views/__init__.py
@@ -1,0 +1,2 @@
+from list import list_bp
+from user import user_bp
